chore: Move useBrowser methods to utils/browser. Replace useBrowser by direct import
