+++ source/components/Header/NormalHeader.tsx
@@ -1,14 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import { Icon, IconButton } from 'components/index';
-import {
-  useStore,
-  useAccount,
-  useUtils,
-  useController,
-  useBrowser,
-} from 'hooks/index';
+import { useStore, useAccount, useUtils, useController } from 'hooks/index';
 import { getHost } from 'utils/index';
 import { Disclosure, Menu, Transition } from '@headlessui/react';
+import { browser } from 'webextension-polyfill-ts';
 
 export const NormalHeader: React.FC = () => {
   const { wallet } = useController();
@@ -16,7 +11,6 @@ export const NormalHeader: React.FC = () => {
   const { activeNetwork, encriptedMnemonic, networks } = useStore();
   const { handleRefresh, navigate } = useUtils();
   const { activeAccount } = useAccount();
-  const { browser } = useBrowser();
 
   const [isConnected, setIsConnected] = useState<boolean>(false);
   const [currentTabURL, setCurrentTabURL] = useState<string>('');
