+++ source/hooks/useAccount.ts
@@ -1,16 +1,14 @@
 import { useSelector } from 'react-redux';
 import { RootState } from 'state/store';
-import { getHost } from 'utils/index';
+import { getHost, getCurrentOrigin } from 'utils/index';
 import IWalletState, { IAccountState } from 'state/wallet/types';
 
-import { useController, useWindowsAPI } from '.';
+import { useController } from '.';
 
 export const useAccount = () => {
   const { accounts, activeAccountId }: IWalletState = useSelector(
     (state: RootState) => state.wallet
   );
-
-  const { getCurrentOrigin } = useWindowsAPI();
 
   const controller = useController();
