+++ source/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './browser';
 export * from './format';
 export * from './getHost';
 export * from './isNft';
